refactor(rust): satisfy clippy -D warnings for Phase 1 gate
Clean-HEAD 'cargo clippy --all-targets -- -D warnings' surfaced 6 lint-as-error
that Phase 1 cycles never ran through. All six are mechanical fixes:

- commands/alerts.rs:40   add #[allow(too_many_arguments)] on add_alert_rule
- commands/server.rs:54   add #[allow(too_many_arguments)] on add_server
- commands/server.rs:187  drop unnecessary `as i64` (get::<i64,_> already returns i64)
- metrics/rollup.rs:61    add #[allow(too_many_arguments)] on generate_rollup
- scheduler/probe_scheduler.rs:133  collapse `} else { if let .. }` -> `} else if let ..`
- scheduler/probe_scheduler.rs:196  same collapse pattern

Three Tauri command signatures are kept as-is via #[allow] because Tauri IPC
DSL constrains arg shape, and packing into a struct would ripple through every
typed wrapper in src/services/tauri.ts plus React callsites — out of scope for
Phase 1 gate fix.

+++ src-tauri/src/commands/server.rs
@@ -51,6 +51,7 @@ pub async fn list_servers(pool: State<'_, SqlitePool>) -> Result<Vec<ServerConfi
 }
 
 #[tauri::command]
+#[allow(clippy::too_many_arguments)]
 pub async fn add_server(
     pool: State<'_, SqlitePool>,
     name: String,
@@ -184,7 +185,7 @@ pub async fn get_server_health_summary(
         for row in &rows {
             let status: String = row.get("status");
             let last_seen_at: Option<i64> = row.get("last_seen_at");
-            let polling_interval_sec = row.get::<i64, _>("polling_interval_sec") as i64;
+            let polling_interval_sec = row.get::<i64, _>("polling_interval_sec");
 
             match status.as_str() {
                 "online" => {
